move some parts to ledger conversion module for better code reuse
